% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vocab.r
\name{get_vocab}
\alias{get_vocab}
\title{get_vocab}
\usage{
get_vocab(train_file, vocab_file, verbose = FALSE)
}
\arguments{
\item{train_file}{TODO}

\item{vocab_file}{TODO}
}
\description{
TODO
}
\details{
TODO
}

