% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2v-package.r
\docType{package}
\name{w2v-package}
\alias{w2v-package}
\title{w2v}
\description{
Bindings for a MPI+threads parallel implementation of word2vec.
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\references{
Programming with Big Data in R Website: \url{https://pbdr.org/}
}
\keyword{Package}

