% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2v.r
\name{w2v}
\alias{w2v}
\title{w2v}
\usage{
w2v(train_file, output_file, vocab_file = NULL, verbose = FALSE)
}
\arguments{
\item{train_file}{}

\item{output_file}{}

\item{vocab_file}{If a string is provided, it will either write 
If \code{NULL}, neither saving nor reading takes place.}
}
\description{
TODO
}
\details{
TODO
}

