% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.r
\name{w2v_params}
\alias{w2v_params}
\title{w2v_params}
\usage{
w2v_params(binary = FALSE, disk = FALSE, negative = 5, iter = 5,
  window = 5, batch_size = 11, min_count = 5, hidden_size = 100,
  min_sync_words = 1024, full_sync_times = 0, alpha = 0.1,
  sample = 0.001, model_sync_period = 0.1)
}
\arguments{
\item{binary}{(boolean) Save the resulting vectors in binary mode?}

\item{disk}{(boolean) Stream text from disk during training? Otherwise the text will be loaded memory before training.}

\item{negative}{(non-negative integer) Number of negative examples. Common values are 3 - 10 (0 = not used).}

\item{iter}{(positive integer) Number of training iterations.}

\item{window}{(non-negative integer) Max skip length between words.}

\item{batch_size}{(positive integer) The batch size used for mini-batch training.}

\item{min_count}{(non-negative integer) Discard words that appear less than \code{min_count} times.}

\item{hidden_size}{(positive integer) Size of word vectors.}

\item{min_sync_words}{(positive integer) Minimal number of words to be synced at each model sync.}

\item{full_sync_times}{(non-negative integer) Enforced full model sync-up times during training.}

\item{alpha}{(float) Initial learning rate.}

\item{sample}{(float) Threshold for word occurrence.}

\item{model_sync_period}{(float) Synchronize model every \code{model_sync_period} seconds.}
}
\value{
A \code{w2v_params} object, for use in \code{w2v()}.
}
\description{
Generates the parameters list for use in \code{w2v()}.
}

