% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vocab.r
\name{get_vocab}
\alias{get_vocab}
\title{get_vocab}
\usage{
get_vocab(train_file, vocab_file, verbose = FALSE)
}
\arguments{
\item{train_file}{Input plaintext file.}

\item{vocab_file}{File path (string) pointing to where you want the vocabulary to be saved.}

\item{verbose}{Want it to print what it's doing?}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Pre-compute the vocabulary for use with \code{w2v()}.
}

