% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2v.r
\name{w2v}
\alias{w2v}
\title{w2v}
\usage{
w2v(params = w2v_params(), train_file, output_file = NULL,
  vocab_file = NULL, nthreads = 4, message_size = 1024, comm = 0,
  verbose = FALSE)
}
\arguments{
\item{params}{A parameters list. See \code{?w2v_params}.}

\item{train_file}{Input plaintext file.}

\item{output_file}{TODO}

\item{vocab_file}{Either a file path (string) pointing to the output of \code{get_vocab()} if
the vocabulary has been pre-computed, or \code{NULL} to compute it on the
fly.}

\item{nthreads}{TODO}

\item{message_size}{MPI message chunk size in MB.}

\item{comm}{An MPI communicator number (from pbdMPI).}

\item{verbose}{Want it to print what it's doing?}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Bindings for MPI+threads implementation of word2vec.
}

